CMake: silence warning configuring examples
When configuring qtwebengine as its own module, Svg is not searched as
an optional component, leading to the warning in
examples/pdf/CMakeLists.txt always being printed even when QtSvg is
available.

Add Svg as an optional package in find_package and move the checks
into the individual examples so the warning will also appear when the
individual example is built.

Pick-to: 6.8 6.9
Change-Id: I387dfd70ae1abc1a70f28dbfe74ac45b75bdf387
Reviewed-by: Michal Klocek <michal.klocek@qt.io>