Fix download behavior when network error occurs
We assume that DownloadManagerDelegateQt::DetermineDownloadTarget() is
always the starting point of a download request, but there are special
cases when a download item needs more than one target determination
(for example when the server is being killed during a large download,
Content API tries to check the target again before reporting the error).
Handle this case by not decreasing m_currentId and preventing the
manager class adding itself as an observer of the same item multiple
times.

Due to the weird internal state of DownloadItem
(INTERRUPTED_TARGET_PENDING_INTERNAL) it still counts as an in progress
download and its functions (e.g. Resume()) are inactive. Don't expose
this state to our users, just re-use the already provided target path
again and they will be informed about the issue in the next
OnDownloadUpdated().

Pick-to: 6.8 6.9
Task-number: QTBUG-132479
Task-number: QTBUG-132473
Change-Id: Id8d39881b9225d76456edec150035e4331e752c7
Reviewed-by: Anu Aliyas <anu.aliyas@qt.io>