Detect and use Widevine libraries shipped with Microsoft Edge
Add detection for Widevine libraries that come by default with a
Windows installation. The libraries may be found both in Microsoft
Edge's installation folder, and also inside the WebView2 data
folders inside the System32 folder.

Also modifies the getters for system folders on Windows, as their
docs state that "SHGetSpecialFolderPath is not supported".

Pick-to: 6.9
Change-Id: I98ebf1f6f801f0ea32be6fd3653c9837f2827592
Reviewed-by: Allan Sandfeld Jensen <allan.jensen@qt.io>