CMake: Handle missing python dependency for Chromium SBOM conversion
By default qtbase always generates tag/value SBOM documents, and only
generates JSON SBOM documents if the required spdx-tools python
dependency is found. If it is not found, it silently skips the json
file generation.

On the other hand, QtWebEngine generates the Chromium specific SBOM
into the json format. We then need to convert that into a tag/value
format using the spdx-tools package, to be able to link the
QtWebEngine SBOM to the Chromium one.

This means that if the spdx-tools package is missing, qtbase only has
tag/value SBOMs, Chromium only has a json SBOM, and we can't link
them. The Chromium one is therefore useless, and there was no reason
to generate it.

Change the build system code in qtwebengine to do the following:
- if the python dependency is missing, skip the Chromium SBOM
  generation and show a configure check warning that the Chromium SBOM
  is skipped
- if the python dependency is missing and
  QT_SBOM_REQUIRE_GENERATE_JSON is ON (or -sbom-json-required is
  passed), skip building both QtWebEngine and QtPdf, because we can't
  satisfy the requirements. The variable is only ON if the user
  explicitly opted into the option. This follows the logic in qtbase.
  Usually we would have errored out in such a case, but we can't due
  to the usual can't error during configure in qtwebengine reasons.

Remove the previous workaround of skipping the Chromium SBOM
generation without showing any diagnostic messages.

This reverts commit 3250fc3cca8b88d6c8ab2f9edc7dd46bfb2d3681

Pick-to: 6.9
Change-Id: I68b7e281a80e8edc11197350b4400408ff0ec8c8
Reviewed-by: Alexey Edelev <alexey.edelev@qt.io>