Fix flaky tst_QWebEngineView::deferredDelete
A replacement widget for the obsolete QDesktopWidget might be created
while hovering a widget in an earlier test during destruction. The extra
widget is destructed together with QApplication so it is listed by
QApplication::allWidgets() in the subsequent tests.

For reproducing this issue run
./tst_qwebengine doNotSendMouseKeyboardEventsWhenDisabled deferredDelete
and keep the mouse cursor above the window while running
tst_QWebEngineView::doNotSendMouseKeybooardEventsWhenDisabled().

As a workaround take the extra widget into account when testing the size
of QApplication::allWidgets().

Pick-to: 6.8 6.9
Change-Id: I84cfb75fcee944e8a22b7c12a725f131e96b3719
Reviewed-by: Moss Heim <moss.heim@qt.io>
Reviewed-by: Allan Sandfeld Jensen <allan.jensen@qt.io>