Enable WoM cross compilation
Support cross compilation for WoA to enable binary releases.
We do not support native compilation as this is way to slow, however
binaries can be released from cross compilation target.

Moreover interrogated change for switching dependency for host_qt
to mingw for msvc22 cross compile target reduces
further the overall integration time.

Note the testing is done on native window arm64.

This setup is a bit hackish but seems to do the job.

Task-number: COIN-1211
Pick-to: 6.9
Change-Id: I0d3a715cad2b69dec8eea66d9e79da8abf705ce1
Reviewed-by: Allan Sandfeld Jensen <allan.jensen@qt.io>