Fix deadlock on WebEngineContext destruction
viz::Display is destructed with the UI compositor
(RenderWidgetHostViewQt::m_uiCompositor). The destruction should be
finished on the Browser thread when cleanupVizProcess() gets called in
WebEngineContext. However, some of the viz::Display members' destructors
(e.g. viz::SkiaOutputSurfaceImpl and
viz::DisplayCompositorMemoryAndTaskController) started destroy tasks on
the GPU thread. These tasks have to be completed before we try to
shutdown threads (BrowserMainLoop::ShutdownThreadsAndCleanUp()) otherwise stopping the VizCompositor thread will cause a deadlock.

As a workaround, post a new task on the GPU thread and wait for it to be
completed.

Pick-to: 6.9
Change-Id: Ie217720008dae8a7b84adb603479b755dd79f5f3
Reviewed-by: Allan Sandfeld Jensen <allan.jensen@qt.io>