Client Hints: Improve version lists
There are several issues affecting Sec-CH-UA and -Full-Version-List,
because the general algorithm which extends and shuffles these lists
was not used in WebEngine.

At the current version of Chromium we can't use the corresponding user
agent utils functions, because they only allow one extra brand/version
pair besides the two generated (Chromium and Not-A-Brand) ones.
This change implements a simplified method to extend and shuffle the
version lists to follow the behavior of Chromium and support more than
three elements.

QTBUG-133799: The order of brands and versions are still not
customizable by the API user, but it is permutated in each major
version of Chromium to comply with the standards. To allow
customization of the order seems to be impossible without changing our
public API.

QTBUG-133708: The missing values of Sec-CH-UA are generated from the
full version list, not customizable individually. This way it keeps
the same values and the same algorithmically generated order as the
full version list.

Task-number: QTBUG-133711
Task-number: QTBUG-133708
Task-number: QTBUG-133777
Task-number: QTBUG-133799
Change-Id: I96f214ce54190666a34779130a04b56f600abef7
Reviewed-by: Peter Varga <pvarga@inf.u-szeged.hu>
Reviewed-by: Moss Heim <moss.heim@qt.io>