tst_MultiPageView::pinchDragPinch: avoid pinching on a PDF link
This test is from c0524da14d92a957b4607fb5867dba5d23eea6d9 Let's assume it was
not intentional to have one of the touchpoints for the pinch gesture actually
landing on a PdfLinkDelegate, which makes the viewer jump to a different page
when it's tapped. qtdeclarative/4a5958f822565bbcb641965ba0f2542f81a42be1
has the effect that the TapHandler is acting as if it was tapped in this case.
So we work around it for now by pinching in a different location.

Task-number: QTBUG-134938
Change-Id: Ia0cf79a1156eea3a56ce0e524a7d063d8e206e24
Reviewed-by: Allan Sandfeld Jensen <allan.jensen@qt.io>