CMake: Fix CMP0177 warnings
INSTALL_DATADIR and INSTALL_ARCHDATADIR have the value "." by default.

CMP0177 in CMake 3.31 now warns when paths that have "." as a
component are passed to install().

To avoid the warning, pass each component of the path separately to
qt_path_join, so that it can properly filter out ".".

This change will have only have an effect once a newer qtbase is
brought in with the updated qt_path_join behavior.

Pick-to: 6.9
Change-Id: Ic8b5e7a1dfe92a16dd7fe69dfddd4141dd648e48
Reviewed-by: Moss Heim <moss.heim@qt.io>
Reviewed-by: Alexey Edelev <alexey.edelev@qt.io>