Create EGLImage with eglCreateDRMImageMESA() for exporting dma_buf
This simplifies the code because it is not needed to create OpenGL
texture to allocate buffer. This way we can get rid of creating a
temporary OpenGL context and swapping EGLSurface.

Moreover, eglCreateImage() does not guarantee that the buffer is
shareable thus certain GPU drivers may fail to export.

Pick-to: 6.9
Fixes: QTBUG-136257
Done-with: Jan Palus
Change-Id: Ie72d052a2a8b1a41bebf0eab8a4928d38c8fc864
Reviewed-by: Allan Sandfeld Jensen <allan.jensen@qt.io>
Reviewed-by: Jan Palus <jpalus@fastmail.com>