Make possible to use system EGL instead of ANGLE on Linux
This is not the same direct EGL rendering solution what we used to have
before QtWebEngine 6.9. With this configuration, the texture is accessed
via NativePixmap instead of using shared GL context. Thus this
implementation doesn't need adaptations in Chromium.

This configuration is meant to be used for identifying ANGLE specific
issues.

Enable it with:
--webEngineArgs --use-gl=egl

Change-Id: Ib8db5573b2001157eaee799c6b333f722ab87604
Reviewed-by: Moss Heim <moss.heim@qt.io>
(cherry picked from commit 18210c2516759d4089a835e384d1b3ef88755c38)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>