Examples: Fix more unqualified lookups in quicknanobrowser
Due to a cyclical dependency error, declare applicationRoot to be a
QtObject. This required all its usages to be as-casted to
ApplicationRoot to avoid warnings. BrowserWindow used in WebAuthDialog
has the same issue. Work around that issue by using parents. This error
is fixed on dev and this these changes are reverted in a following
commit targetting only dev.

Task-number: QTBUG-138589
Change-Id: I7327ec298432b1cfbb6d34f994b41cff0ab19eab
Reviewed-by: Sami Shalayel <sami.shalayel@qt.io>
(cherry picked from commit 1059b682758323a3527a5edb0cd7477e0827be73)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>